+++ tests/failures.nix
@@ -1,5 +1,6 @@
 {
   crossConfig,
+  flakeParts,
   mkNodes,
   nixpkgs,
 }:
@@ -158,3 +159,4 @@ in
 // import ./destination-failures.nix { inherit mkNodes; }
 // import ./tagged-failures.nix { inherit mkNodes; }
 // import ./module-failures.nix { inherit crossConfig nixpkgs; }
+// import ./flake-module-failures.nix { inherit crossConfig flakeParts; }